Fixed issue with tests failing to run
