+++ src/embedDB/embedDB.h
@@ -254,6 +254,13 @@ typedef struct {
  */
 int8_t embedDBInit(embedDBState *state, size_t indexMaxError);
 
+/* Constructors */
+/**
+ * @brief	Initialize embedDB structure with default parameters.
+ * @return	Returns an embedDB state with default values if successful, and NULL if not
+ */
+embedDBState *defaultInitializedState();
+
 /**
  * @brief   Prints the initialization stats of the given embedDB state
  * @param   state   embedDB state structure
